DROID-2493 rename
