+++ payments/src/main/java/com/anytypeio/anytype/payments/mapping/MembershipResourcesExt.kt
@@ -3,30 +3,10 @@ package com.anytypeio.anytype.payments.mapping
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.pluralStringResource
 import com.anytypeio.anytype.core_ui.R
-import com.anytypeio.anytype.payments.constants.TiersConstants
 import com.anytypeio.anytype.payments.models.PeriodDescription
 import com.anytypeio.anytype.payments.models.PeriodUnit
 import com.anytypeio.anytype.payments.models.TierPeriod
-import com.anytypeio.anytype.presentation.membership.models.TierId
 import java.time.Period
-
-fun TierId.getTierTitle(): Int {
-    return when (this.value) {
-        TiersConstants.EXPLORER_ID -> R.string.payments_tier_explorer
-        TiersConstants.BUILDER_ID -> R.string.payments_tier_builder
-        TiersConstants.CO_CREATOR_ID -> R.string.payments_tier_cocreator
-        else -> R.string.payments_tier_custom
-    }
-}
-
-fun TierId.getTierSubtitle(): Int {
-    return when (value) {
-        TiersConstants.EXPLORER_ID -> R.string.payments_tier_explorer_description
-        TiersConstants.BUILDER_ID -> R.string.payments_tier_builder_description
-        TiersConstants.CO_CREATOR_ID -> R.string.payments_tier_cocreator_description
-        else -> R.string.payments_tier_custom_description
-    }
-}
 
 fun String.parsePeriod(): PeriodDescription? {
     return try {
